tests: Remove PATH management code for nvdisasm
The test suite (naturally) had a bunch of code designed to try and
detect when `nvdisasm` needed to be on the PATH or not, since the
behavior of `hpcstruct` would change drastically if it was missing.

Now that the behavior of `hpcstruct` can be trivially predicted
beforehand (and mistakes result in command failure instead of other
bizarre errors), remove this thorny cruft code and greatly simplify the
`hpcstruct` tests.