refactor(dev): Remove the duplication of Spack specs
Remove some of the duplication of information by reading the core
dependencies from the spack.yaml files used for the subprojects, and
use the subproject fallbacks instead of native file settings to provide
those dependencies.

(Almost) breaking change: This commit causes the `./dev` script to
depend on Spack outside of Spiqa. Spiqa currently hard-depends on Spack
in the same cases, so this does not actually change much, but just FYI.