ci: Move CI image management code to this repo
Currently, the images used in the CI pipelines here are generated by the
separate https://gitlab.com/hpctoolkit/hpctoolkit-deps repo. While this
is efficient, it causes significant headaches when the image build
scripts are adjusted, e.g. release branches should use the older builds
while `develop` needs the newer. There are version numbers to make this
"work," but it's not convenient nor intuitive, and very easy to
accidentally abuse.

The original intention for this split was to ensure as little time as
possible is wasted building images. However, it has not been intuitive
for developers to manually fire off the pipelines to do this, and these
pipelines have been required more often of late.

This commit moves the build scripts (i.e. Containerfiles) used for the
CI images back into this repository. These images are cached and rebuilt
on an as-needed basis, using the CI component/template now present in
https://gitlab.com/blue42u/ci.predeps.

The biggest benefits from this approach come from the ability to freely
customize the build scripts. Custom OS packages (e.g. `ccache`) can be
pre-installed and Meson native files pre-generated for various software.
These new benefits will be exploited in the following commits.