ci: Make use of dependencies installed in ci.predeps images
Now that we can customize the iamges used in CI without minimal
overhead, start exploiting this in our CI jobs by removing the pieces
that previously were only possible with the `./dev` scriptsuite.

The build scripts added in the parent commit pregenerate a Meson native
file with all the software included in the image, `image.ini`, as well
as a modern copy of Meson. This removes the need for `./dev populate`
and `./dev meson`, so those commands have been replaced by their
original `meson` counterparts.

The calls to `./dev nproc` in CI have been removed. It never was
required on the GitLab shared Runner fleet, and planned improvements to
the Rice Runner infrastructure should (eventually) make it unnecessary
there too.

The remaining calls to the `./dev` script (`./dev cc-diagnostics` and
`./dev validate-install`) have been commented out. The missing
functionality will be reintroduced in better homes in later commits.