Fix header and path setup in configure
The way that Red Hat lays out files within the distro was making them
so that they were not being found by the configure script. I hacked in
some logic to find the files in the places where they are in Red Hat
like distros.

One problem with libiberty required changing the name of a header file
so that it wouldn't be found before the system demangle.h in the
search path. This of course required changes within all of its users.

Signed-off-by: Ben Woodard <woodard@redhat.com>