ci: Drop install tree validation (validate-install)
There is very little practical benefit that comes from running
validate-install in CI, the most that happens is a CI failure whenever
you try to add new functions to libhpcrun.so. The manifest itself is
both fragile and complex, and is already listed in the build system
itself.

Given everything, let's just drop this check and move on.