build(meson): Remove Autotools version restrictions
The generated Autotools files are no longer being tracked in Git thanks
to the previous commits, so the version(s) of the Autotools used during
the build no longer needs to be consistent between multiple builds.

Remove the version constraints enforced at the Meson layer and use
whatever Autotools are available.