refactor: Remove non-standard stdatomic.h
C11 adds standardized atomic operations in the header `<stdatomic.h>`.
Historically, we've had a copy of this header in our source tree stolen
from an old FreeBSD, that translates the C11 syntax into the older GCC
extension for the same. This header is incompatible with the standard
header of the same name, which can cause compile errors if both are
included in the same compilation unit.

Our supported compilers all support C11, so remove the old and
unnecessary header and use the system stdatomic.h.