fix(hpcrun): Support compiling (some) headers under C++
There's a fragment of C++ code in hpcrun which includes a number of
internal headers. With the shift to the C11 `stdatomic.h`, this fact now
exposes a completely new hiccup: many of the headers use C11 atomics but
C++ does not have `stdatomic.h` (before C++23).

(A better way to interpret the situation is as follows: C and C++ are
two separate languages, but a subset of each *happens* to be
syntactically, semantically and ABI-identicial when interpreted under
both. This subset did not include C11 atomics until C++23.)

This commit goes through the compilation failures and replaces the
affected code with the C++ `<atomic>` equivalent when compiling as C++.
For most compilers (GCC included), C11's `_Atomic T` and C++11's
`std::atomic<T>` are ABI-equivalent, so these changes are effectively
syntactic at most.

Note that this is not a comprehensive overhaul to ensure every header
can be compiled under C++, this is an as-needed modification to make the
compiler happy.