refactor!: Remove the ./dev script-suite
There is no longer a reason to keep the `./dev` script-suite around, its
functionality has been successfully subsumed into the rest of the
codebase or replaced with cleaner approaches, such as direct-to-Meson
builds. It never collected many users (just me and the CI), the users it
had have transitioned to Meson builds, and no new users will be using
the `./dev` script for their builds. It is time to sweep up and move on.

This commit deletes the `./dev` script-suite from the codebase.

Implements https://gitlab.com/hpctoolkit/hpctoolkit/-/issues/807