build(meson): Fix build when using Spack-built Autotools
If `autoconf` and `libtool` aren't installed to the same install prefix,
the `libtool` commands in `configure.ac` will fail to find the proper
M4 fragments needed to use `libtool`, failing the build. This can happen
when the Autotools are installed via Spack and aren't installed on the
base system.

When building under Meson, run `libtoolize` explicitly to install the
`libtool` files into the source tree before running `autoconf`. This
avoids the need for Meson to find the libtool M4 sources and extend
`ACLOCAL_PATH` to include them.

Note that this extra call is only added to `meson/autogen` and not to
`./autogen`. The assumption is that those building via Autotools (i.e.
Spack, for now) will either have a consistent install prefix for all of
the Autotools or will adjust `ACLOCAL_PATH` accordingly.