fix: Minor fixes to .gitignore paths
`configure` and `aclocal.m4` should only be ignored if they appear in
the top-level source directory, and not if they appear in lower
directories. Fix this to avoid matching the `meson/configure` script.

Also ignore `~`-suffixed backup files generated during `autoreconf`
when it makes changes, and the entire `autom4te.cache`.