refactor(hpcstruct): Remove wrapper script
Previously this wrapper script added the install prefix's `ext-libs/`,
but it also set `OMP_STACKSIZE` based on a `-s/--stack` argument passed
to the script (or a default). Sometimes this is needed to make Dyninst's
various recursive algorithms work.

This commit reimplements the management of `OMP_STACKSIZE` in the
`hpcstruct` executable itself, it will now re-exec itself if it needs to
change the value. This in addition to the prior removal of `ext-libs/`
makes the wrapper script a no-op and it is safely removed.