hpcrun: Corral symbol interceptions into foils
We intercept a lot of symbols, and many symbol interceptions (or just
exposed symbols) are implemented in different ways using different
macros for ultimately the same task. This commit refactors them all to
use a single consistent pattern, the "foil" pattern.

The foil pattern consists of two functions. The "foil(s)" are small
functions that handle symbol interception, live in application-space and
whose primary goal is to call the corresponding "base". The "base" for a
foil is an internal function that reacts to the interception, it has
full access to hpcrun functions and usually calls the "real" function
passed to it from the foil.

The foil(s) and base may be in different `.so` files, usually the foil
is in a "plugin" library (e.g. `libhpcrun_opencl.so`) and the base is in
the main library (`libhpcrun.so`). Foils access their base functions via
one of the few HPCToolkit symbols exposed to the application,
`hpcrun_foil_base_lookup`. Calling this with the symbol name of a foil
returns a pointer to the corresponding base function. For simplicity,
the base functions are all internally named `foilbase_<symbol name>`.

There may be multiple foils for a base function, for instance using
LD_PRELOAD (`*-preload.c`) or `ld --wrap` (`*-wrap.c`). The main goal
for this refactor is to allow alternative interception implementations.
The code for all foils is stored in the `foil/` subdirectory.

Implements https://gitlab.com/hpctoolkit/hpctoolkit/-/issues/792