hpcrun: Hide symbols using visibility
The current hpcrun link uses `objcopy` to try and limit the symbols that
`libhpcrun.so` exposes. This is nonstandard and just plain inconvenient,
so reimplement using GCC-style function visibility attributes instead.

All code is now compiled with `-fvisiblity=hidden`, which means
function symbols will not be exposed after the link unless they are
explicitly marked with `__attribute__((visibility("default")))`. This
puts the documentation for what symbols are exposed *on the functions
themselves* instead of in a confusing nonstandard fragment of Autogoo.
This also means the `libhpcrun.so` in the build directory is the
"completed" version that gets installed, there is no final link during
the install.

As it happens the only symbols that are exposed now are foils, or
internal symbols used for the private namespace. This is entirely
intentional and comes out rather convenient.