refactor: Support building/testing with Python 3.8
Most of the Python code in the codebase is written for Python 3.10+,
including parts of the build system and `hpctesttool`. It is relatively
easy to get a newer Python version, but since the build system only
allows "one" Python for both the test suite and Python logical unwinding
support (`-Dpython=enabled`), this may cause unnecessary friction when
building on a system far afield.

This commit smooths the case where you don't need Python logical
unwinding support, it's a fairly large refactor to downgrade the Python
used in the test suite to Python 3.8. You may now build, install and
test HPCToolkit via Meson with Pythons as old as 3.8.

Note that the Python logical unwinding support (`-Dpython=enabled`)
as of this commit still requires Python 3.10. This commit permits
further investigation into whether this restriction can be loosened in
the future.