refactor: Untrack all Git LFS files
Git LFS is a giant stumbling block for users and developers alike, and
the benefits to using it are in general minor. Move all the files
currently in LFS over to Git proper.

Some significant reorganization of the test data files is done here to
take advantage of Git's delta compression (and to just simplify the
build and test system in general). Specifically:

- All tarballs have been unpacked. It turns out a sizable amount of the
  data can be reused (e.g. same Structfiles), so the Git deltas that
  result are tiny when the files are unpacked.

- All non-essential files in measurement directories and database
  directories has been pruned out. That means `*.log`, leftover cruft
  from the `hpcstruct` `make`, `FORMATS.md` and `METRICS.yaml.ex`.

- Format version numbers have been removed from the tree. In practice we
  care much more about the "current" format of the files than older
  formats, at least for the component regression testing we are limited
  to today. And these version numbers were always confusing to maintain.

- The struct/ data has had its `small/` prefix removed. They're all
  "small" enough to use for testing, so having the extra prefix just
  makes things unnecessarily complicated.

- All *.hpcstruct files are marked so that Git treats them as binary
  instead of as text. This avoids Git generating massive textual diffs
  when the changes are complex, as they get in Structfiles.

This is kept as a separate commit to facilitate review, the following
commit repairs the build and test system to work with this organization.

Implements https://gitlab.com/hpctoolkit/hpctoolkit/-/issues/809