build(tests2): Reorganize the test data
This commit is the build and test system updates needed to use the new
organization for the test data.

The biggest change is how test data is regenerated: instead of producing
a `fresh-testdata-X.tar.xz` tarball, there are now `regen-testdata-X`
targets that will regenerate the test data and copy it to the source
directory in one step. It's done this way because there are files in
the measurements directories that change names on regeneration, which
won't be exposed when expanding a tarball.