build(tests2): Adapt to new locations for test binaries
Now that test binaries are in this repository, simplify the code
surrounding their management in the build system. This doesn't change
the end result of the tests but does remove a number of conditionals.

One major advantage is that the listing of the test binaries is now in
Meson instead of keyval, which opens up some more options for how the
data is used and written. So benefits there.