doc/man: Write man pages in reStructuredText
Currently our man pages are written in a dialect of LaTex and then
translated (via custom LaTeX packages and ancient Perl scripts) into
man pages. In addition to the technical issues with this arrangement,
as a markup language LaTeX has a **very** steep learning curve and is
overly heavyweight for this task.

This commit transcribes the man pages to reStructureText (rST). rST is
chosen here as it is a lightweight markup language with minimal learning
curve, widespread tool support and standard support for syntactic
features common in man pages: authorship and other metadata, option
lists, definition lists, and admonitions.

This commit tries to retain the *content* of the text, but dramatically
reformats the rST produced by Pandoc to produce clean and writable rST
documents. This makes editing the man pages accessible for many, in the
spirit of [#716].

Significant efforts are also made to ensure the man pages as rendered by
`rst2man` are typeset appropriately following man page conventions. In
particular, keywords are typeset as code (bold) while user-defined
variables are typeset in italics/underline, and definition lists are
typeset without the leading bullets.

[#716]: https://gitlab.com/hpctoolkit/hpctoolkit/-/issues/716