test(python): Allow "gaps" above Python unwind
It's possible `python3` is a separate binary from `libpython.so`, in
which case the Python logical unwinder is unable to remove *all* frames
above the Python code even if the main executable is Python.

Adjust the test to allow up to 4 "gap" frames between the entry point
and Python logical contexts. This fixes the tests on ppc64le.