fix: crash when receiving Python notifications without a logical stack
Previously out of an abundance of caution, the Python unwinder would
`hpcrun_terminate()` whenever an attempt to return from a function or
call into C from Python was encountered but there was no logical stack.
This is fine for testing but not for experimentation.

Alter the unwinder to just ignore such notifications if they occur. The
only case they could occur in the wild is if an application shares a
Python "thread" between CPU threads. So far it seems rare for
applications to ever actually do that.

If that ever does happen, there *should* be enough robustness in the
Python on-demand unwinding and common logical code to at least not
crash. Probably.