fix(hpcrun/python): Remove incorrect complaints about missing frames
When Python calls C, we inject an extra logical "C function" frame above
all the Python frames to record the Python name for the C function being
called. Unfortunately the detection logic for missing frames did not
take this into account, and so would always report "some missing Python
frames" when in fact there was no problem.

Fix the logic for this warning so it properly counts the C function
frame.