ci: Simplify the spack install test
Instead of using a predeps image, use a normal upstream Spack image and
install HPCToolkit by pulling from the (now OCI) buildcache. This
removes the difficulty with trying to merge the Spack environment used
in CI and what Spack develop wants, allowing the two to diverge as
needed.

This may make the job run longer, since now Spack is allowed to choose
not to reuse packages. In practice this should be rare, and building
fully from scratch is not *that* slow.