build(predeps): Support --layers
Previously the ci/predeps/ container scripts used a number of bind-mount
tricks to ensure unwanted cruft would not be left in the final image.
This unfortunately meant if the image was built with `--layers` the
layer cache would not be busted when needed thus causing mayhem.

Rewrite the script to use `COPY` commands which do bust the cache, and
then use a multi-stage build to ensure the generation scripts don't
appear in the final image.