build(predeps): Remove the REMOTE_CACHE_MIRROR build-arg
Instead of specifying the mirror to push to as a build argument,
hard-code an obviously special name that no-one will have as the name
for a mirror they actually care about.

This removes an unnecessary layer cache miss when only the only change
is in the secrets, not the actual content.