Remove non-default -I directories
The build (esp. for hpcrun) includes a number of -I flags that refer to
various directories in the source tree. This makes it non-intuitive to
know how to write an `#include`, for example should `messages.h` be
included as `messages.h` or as `messages/messages.h` or as
`tool/hpcrun/messages/messages.h`?

To make matters worse, the unwinder code is segmented using conditional
include directories, meaning a number of headers with identical names
are `#include`'d by name. This is confusing and makes the build system
and code remarkably brittle.

This commit removes these flags from the build scripts.