Use relative #includes
This commit reformats all the #includes in the codebase, such that all
files that are in the codebase are #included by a relative path to the
file. This makes it very clear which file is being #included and removes
the assistance of e.g. a compilation database to know for certain.

In addition, all includes that refer to our code are consistently
reformatted to use the local syntax `#include "..."`, while all
includes that refer to others' use the system syntax `#include <...>`.
This makes it clear where to go look for a header, as well as permits
using `-isystem` for some purposes.

As a special case, certain generated headers (`hpctoolkit-config.h`) are
referred to by name (e.g. `#include "hpctoolkit-config.h"`). Autotools
and Meson both support and encourage this syntax.