build(predeps): Freeze at Spack 0.21.0
CI needs to be extremely reliable and reproducible. "Rolling" software
is the cause behind the worst failures, since a behavior or config
format change causes the CI to fail suddenly and without warning,
bricking all MRs until the issue can be fixed upstream.

To avoid as much of this as possible, we freeze the version of Spack
we build everything with. The best target is a release, so we freeze
Spack to the recent 0.21.0 release.