refactor(hpcrun)!: Remove the libmemkind path override
Currently we tweak `LD_LIBRARY_PATH` to provide a "custom" `libmemkind`
to the application. This approach has many flaws, goes against our
general "don't affect the application" principle of `hpcrun`, and blocks
the Meson port (since it requires a full install path).

Remove the override, and with it the `hpcrun --default-memkind` option
and the dependency on `memkind`.

Resolves https://gitlab.com/hpctoolkit/hpctoolkit/-/issues/810