ci: Allow running check jobs in VMs
Now that we're not using hardware perf events, the test suite should be
able to run to completion inside a VM. Remove the `*-bare-*` tag clause
to represent this shift in requirements.

Note that we still require a sufficiently low value for
`kernel.perf_event_paranoid`, so if the value is too high try to lower
it before running the tests.