ci: Wipe builddir during check semifresh
Currently test data is regenerated into the build directory before being
copied into the source directory. For `check semifresh`, that means the
original file paths for measurement directories still exist, which can
change the result of `hpcprof`.

To avoid this, `--wipe` the build directory after generating the test
data and before testing with it.