test(data): Remove loops-hip-amd-t.d
All AMD measurements (i.e. loops-hip-amd-t.m) contain a "logical" file
in them. Unfortunately that means they can't be moved to a different
location, otherwise `hpcprof` will be unable to find the logical file
and instead produce weird instruction contexts. This issue is deeply
related to https://gitlab.com/hpctoolkit/hpctoolkit/-/issues/411.

As such, this database is always out of date and unable to be used for
regression testing. Remove it.