Replace xed_operand_values_get_branch_displacement_int32() with _int64.
XED used _int32 up to version 2023.07.09 and then switched to _int64
starting with 2023.08.21.  Using the wrong version of XED can cause
hpcrun to sometimes fail with:

   symbol lookup error: .../libhpcrun.so:
   undefined symbol: xed_operand_values_get_branch_displacement_int32

Note: this is a hard switch in XED, there is no overlap.  Hpctoolkit
on x86 up to 2023.08.1 must use XED 2023.07.09 (or earlier) and
toolkit develop (and later revs) must use XED 2023.08.21 or later.