Refactor the ci.predeps container build scripts
The content has not changed, but the code has been reorganized to reduce
the number of build arguments needed to build any of the images. This
means each Containerfile now is written for one--and only one--base
container image. Any OS-specific code is now in the Containerfile, any
generic code is in the extra scripts.

For similar reasons, the image.ini files are now separated by their OS
instead of being generated from patches. This makes them easier to edit
and more consistent anyway.

For Ubuntu: Multilib paths are symlinked away to emulate the RHEL install
paths and make ./configure happy. This will hopefully be temporary.