Use OS-packaged build tools in CI
Most of the build tools we use to build HPCToolkit don't *need* to be
installed via Spack, and there isn't much difference depending on which
version is used. For example, the Autotools and Python.

Instead of installing these tools via Spack, install them using the
packages distributed by the OS.