hpcrun: Abstraction for a private dlmopen'd namespace
Despite our best efforts, we sometimes interact poorly with vendor
libraries or applications, causing crashes, deadlocks, etc. Avoiding
this tail-stepping requires great care and a few hacks on our part to
"hide" ourselves from the application/libraries. Private linkage
namespaces provide an alternate (and much stronger) mechanism to achieve
this "hiding" without nearly as much effort on our part.

This commit adds a private linkage namespace dedicated to `hpcrun`
internals. The namespace itself is either created with `dlmopen()`, or
is the namespace opened automatically for the auditor.

The namespace is accessed through a generic and intentionally simplistic
`hpcrun_bind_private()` function, which takes the library name and a
series of symbol name -> function pointer pairs. The library is loaded
into the private namespace, and the symbols fetched and bound to their
respective locations.

For completeness, there is also an `hpcrun_bind()` function that acts on
the main application namespace instead. This abstraction is intended to
simplify the refactors necessary when `libhpcrun.so` is moved into the
private namespace wholesale.