hpcrun: Move libunwind into the private namespace
`libunwind` is a delicate library that we run under a signal handler.
To protect it from outside wrappers and such, put it in the private
namespace.

More specifically, this adds an `libhpcrun_unwind.so` shim library that
exposes the functions from `libunwind`, those are then bound to a
corresponding series of `libunwind_*` function pointers which are used
in `libhpcrun.so`.

Co-authored-by: John Mellor-Crummey <johnmc@rice.edu>