hpcrun: Remove hpcrun_real_(dl_iterate_phdr|mmap)
These functions are the overriding implementations of `dl_iterate_phdr`
and `mmap/munmap` specifically for libunwind. Now that libunwind is in
a private namespace, we can redesign the overrides to use it instead:

- `mmap/munmap` already will bind to the "true" implementations in libc,
  so we don't need the overrides for them at all.
- `dl_iterate_phdr` can be overridden by defining it in the "head" of
  private namespace, which happens to be where `hpcrun_iterate_loadmap`
  is exposed to the private namespace today.

This commit disables the symbol renaming for libunwind and liblzma is
disabled, the objcopy code remains temporarily as a no-op.