Remove the objcopy symbol override hacks
Now that the libunwind symbol overrides are handled using a private
namespace, there is no longer a need to use objcopy to rename symbols.

This commit removes the objcopy hacks, and then also changes the private
namespace shim to link against `libunwind.so` instead. This removes the
need for an -fPIC'd `libunwind.a`, which allows us to use more typical
libunwind builds and removes the `+pic` dependency in Spack.