Fix setup failure when mpiexec is not found
When mpiexec is not available, it gets replaced by a shell command that
results in an exit code that gets recognized as a SKIP (77). But Meson
quirkily doesn't allow passing the result of `find_program()` as part of
`test(args:)`. To make it work, we use the `full_path()` instead.