hpcprof: Replace stable_hash implementation with XXHash
The stable_hash implementation used a hacky variant on MD5 for the hash.
Apparently, it doesn't work quite right. It could be fixed, but it would
be better to use a real hash function instead.

XXHash is extremely fast and good. Let's use it.

For convenience, a wrap for XXHash is included.