hpcrun: Fix a misplaced lzma link arg
Under Autotools (libtool), the linker is especially sensitive to the
order in which link flags are passed on the command line. Symbols can
only be provided by objects listed "later" on the line, and LDFLAGS come
before LIBADD. So if libhpcrun.so doesn't *directly* call any `lzma_*`
routines but prof-lean or support-lean do, the link will not link
liblzma and produce errors at runtime.

Fix this by moving the lzma link args to LIBADD instead.