Remove compilation command lines from Meson build log
Autotools (technically, libtool) is extremely noisy when building by
default, the entire command line for each compilation step is printed
verbatim to the output log. This noise makes it difficult to find errors
or other extremely important tasks. Instruct libtool to remove this
noise and build "silently."

This new behavior can be overridden by setting `V=1` in the environment
when compiling.