Support running tools from the build directory
When developing, it's a *lot* easier to work with tools that aren't actually installed in their
final location, i.e. from the build directory. Saves a call to `meson install`. `hpcprof` already
has this property, so this commit brings `hpcrun` up to parity (so long as HPCTOOLKIT_FNBOUNDS_CMD
is preset in the environment).

`hpcstruct` almost has this property, but the Makefile-based parallelism strategy for processing
measurement directories makes this far more difficult that it needs to be. For now, the user must
set HPCTOOLKIT_HPCPROFLM and HPCTOOLKIT_HPCSTRUCT for this use case to work properly. The single-
binary use case works with no change to the environment.

It should be noted that the use cases noted above that require additional environment variables
have been slated for reimplementation for multiple years now, `hpcfnbounds` is to be integrated
into `libhpcrun.so` and `hpcproflm` is unnecessary with a non-make-powered parallelism strategy.

Closes https://gitlab.com/hpctoolkit/hpctoolkit/-/issues/781