Test the tools in the build tree
...instead of the tools that have been installed. This makes it easier to quickly test changes
during development, and matches the standard Meson compile-test-install pattern.

The README has been updated to remove the notice of deviation from standard Meson practice, and the
CI now no longer installs before testing as is appropriate.

For convenience, the environment modifications needed for the tools to work when running from the
build tree are set under `meson devenv`, as well as the `PATH` extended to include the tools.
You're welcome.