Check for defines with ifdef
The syntax `#if VAR` only works if `VAR` is defined as a value, say, `1`. Meson by default doesn't
define `VAR` with any value, so the above syntax fails to compile. Replace the few uses of the
above with `#ifdef VAR`, which is semantically equivalent under Autotools and behaves nicer for
Meson.