hpcrun: Remove references to dependency install prefixes
Meson doesn't have install prefixes for dependencies. The only place
the `hpcrun` launch script uses them is to fill in `LD_LIBRARY_PATH`,
but there should already be R*PATHs in libhpcrun.so for that sort of
thing. So remove them.

The only unique case is libmonitor.so, which needs to be listed in
`LD_PRELOAD` as well. This one is worked around by using `ldd` on the
main libhpcrun.so, which links against libmonitor.so. This weirdness
can be removed once [#793] is implemented.

[#793]: https://gitlab.com/hpctoolkit/hpctoolkit/-/issues/793