tests: Increase priority of serial tests
The order in which tests are launched happens to be the same order they are
specified in the `meson.build` files, so if a parallelizable test (`is_parallel: true`)
is sandwiched between two serial tests (`is_parallel: false`), then the tests
will run in serial and not take full advantage of the available hardware.

This commit increases the priority of serial (`is_parallel: false`) tests
to 100, ensuring that they start (and finish) before any of the parallelizable
tests start. This allows `meson test` to run as many tests as possible in parallel
and take at least some advantage of the available hardware.