Add devcontainer support
Devcontainers are a VSCode extension that allows developing in
containers instead of on the base system. The goal is to provide (many)
reproducible software environments that can be shared among developers
and easily spun up to test particular features.
See https://containers.dev/ for more information on the project.

This commit adds devcontainers that are identical to those run in CI.