devcontainers: Add VSCode customizations
VSCode has an extension for IDE integration with Meson. Configure the
devcontainers to install this extension, and also configure a Meson
build directory tuned for the image being used.

To keep them out of the way, the build directories are created as:

    .devcontainer/<spin>/builddir-<hash>

This format is chosen as it is least likely to conflict with the build
directory on the host -- or anything else, for that matter.
Meson automatically generates a .gitignore in every build directory, so
this directory will never be committed by accident.