Add framework for using Meson wraps
Meson wraps are a convenient way to provide dependency fallbacks, when
they aren't already available on the system. This commit adds the
skeleton necessary to start using them for real.

CI will not use wraps in the build jobs, except in the case of the new
`bare` predeps image. This image is intended to provide the bare minimum
software necessary to obtain a working HPCToolkit, anything that it
doesn't provide must be obtained by wraps. As wraps are added, the
dependencies can be removed from this image to test them in CI, as part
of the new `check: [bare]` job.