Add a wrap for GTPin
GTPin is a complete snowflake when it comes to software dependencies, it
only comes in binary tarballs and can't be found via CMake due to the
use of non-standard (non-FHS) paths. This means users must expose it to
Meson via `*_args` options... or, maybe, load a module. I don't know. 🤷

Since this is so much trouble to deal with, if `-Dgtpin=enabled` but
isn't available to the compiler, download the public release ourselves
via the Meson wrap system.

For simplicity and consistency, CI will only ever use the wrapped form
of GTPin, which is currently GTPin 3.7.