Refine support for Dyninst master (>=13.x)
Previously attempting to build with Dyninst master would result in an
error in `meson setup`, however the suggestion of adding `modules:`
breaks support for Dyninst 12.3. This commit overhauls both the Dyninst
dependency() calls and the Find script to support both Dyninst 12.3 and
later.

A number of the compiler checks later for Dyninst features also did not
work, mostly because Meson's implementation of various header-symbol
checks is designed for C and not C++. Replace those that have problems
with explicit compiles() checks crafted for C++.

Also fix a missing dependency in `lib/banal` on Xed, which only appears
when using Dyninst master.