Add a wrap for Dyninst
Dyninst is only packaged for Fedora/RHEL. Provide a wrap to build it
from source on Debian/Ubuntu when it isn't available otherwise.

Since we're using the wrap now, remove Dyninst from the ci.predeps images
and always build it from the wrap in CI (except during `spack install`).
We currently need a newer version of Dyninst than is packaged for the
distros so this is the appropriate design as per
https://gitlab.com/hpctoolkit/hpctoolkit/-/issues/812.