hpcprof-mpi: Pass proper types to MPI_* functions
MPI_Gather and MPI_Scatter take extra arguments for the type
received/sent by the call, the MPI standard marks them as "significant
only at root." So when we're not the root rank, we expect the
implementation to ignore such arguments and just pass `nullptr`.

Apparently though, in some environments late-model OpenMPI will in fact
try to dereference the `nullptr` and rapidly segfault. Seems silly, but
this commit makes the calls pass in the same type twice to avoid this.

I hate MPI.