hpcprof-mpi: Add an explicit MPI barrier
OpenMPI is notorious for not allowing parallel Send/Recv pairs to
complete while in a collective. This can cause deadlocks if the flush
on a rank requires a message to rank 0 but rank 0 has already moved on
into the collectives. I hate MPI.

Add an explicit barrier to ensure all the Send/Recv pairs have happened
before entering any of the collectives.

Effect on performance not tested.