ci.predeps: Add OpenSUSE Leap 15 (leap15)
SUSE Leap 15 is the upstream for CrayOS, which some supercomputers run.
It has unique quirks not seen on any other OS to date, such as
https://gitlab.com/hpctoolkit/hpctoolkit/-/merge_requests/1018.

Add a ci.predeps container image for it. Note that there are a number of
slight differences to make this work, e.g. gcc-12 is the default instead
of gcc (which is 7.x).