Localize changes from VCS changes
Ninja is extremely intelligent when compiling, if a header did not
actually change *in content* dependent builds will not fire.
Currently changing branches or making commits triggers a massive rebuild,
all because of `HPCTOOLKIT_GIT_VERSION` which is defined in
`hpctoolkit-config.h`... and only used in `hpctoolkit-version.h`.
`ccache` can speed this up, but it still requires a preprocessor match.

Move the `vcs_tag()` from the generic `hpctoolkit-config.h` to the
specific `hpctoolkit-version.h`, ensuring only a few files (~7 commands)
are rebuilt on VCS changes.