ci: Remodel the pipeline from scratch
The big idea here is to move away from "`buildmany`" jobs and towards
jobs that actually run tests. Why? "buildmany" jobs are useful for
enumerating a huge matrix of environments, but just building usually
isn't sufficient to say our tools actually work.

A more complete solution is to run the tests too, but the huge matrix
enumerated in "`buildmany`" is too large to practically run tests for.
So, instead, we opt to simplify the matrix by focusing on particular
proxy environments and configurations instead of the entire matrix.

The jobs are grouped into stages based on the kinds of things that are
being tested, for instance testing in the CUDA 11.6 container is an
"individual test" while testing that `-Dopencl=disabled` doesn't break
anything is a "configuration test." The classification is very general
and doesn't affect the parallelism (we use needs: now).