tests: Increase test-level parallelism
A number of the tests in the regression test suite are multithreaded,
which increases the overall burden of the test on the underlying system.
Conservatively, we've set these tests so that only one runs at a time.

Be a little more aggressive, and allow running multiple multithreaded
tests at a time... but only up to 3 or 4 threads. The tests generally
finish in seconds, so they should still complete under the 30 timeout
if everything goes smoothly.

Note that we can't in general be this aggressive with MPI, as some
implementations react funny when we try to run multiple jobs on
overlapping hardware. Sigh.