README: Add an explicit note not to install
The way the build environment is produced from the Spack environment,
it is difficult to generate a usable install tree. This will be ironed
out as we move towards system software and wraps, but until then be
louder about telling folks not to `meson install` and to instead use
`meson devenv` for most tasks.